Move Viewport to helpers (#22)
