+++ app/helpers/viewport.rb
